Versão inicial da API 0.0.1

+++ ProvaML.Migration/Scripts/0001_Create_table_Produtos.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE Produtos
 (
-    ID              smallint        identity       primary key,
+    ID              int        identity       primary key,
     Nome            varchar(255)    not null,
     ValorVenda      numeric(15,2)   not null,
     Imagem          varchar(255)    not null
